chore: Comment out audio playback in Home component
The code changes comment out the audio playback in the Home component. This is done by commenting out the creation of the `Audio` object and the `audio.play()` method. This change is made to temporarily disable the audio playback functionality.

Note: The recent user commits and recent repository commits do not provide any additional relevant information for the commit message.